Add reports app and begin refactoring report form into it
